+++ src/plot/layer/geom/range.rs
@@ -44,6 +44,10 @@ impl GeomTrait for Range {
         ];
         PARAMS
     }
+
+    fn supports_aggregate(&self) -> bool {
+        true
+    }
 }
 
 impl std::fmt::Display for Range {