+++ src/plot/layer/geom/segment.rs
@@ -38,6 +38,10 @@ impl GeomTrait for Segment {
         }];
         PARAMS
     }
+
+    fn supports_aggregate(&self) -> bool {
+        true
+    }
 }
 
 impl std::fmt::Display for Segment {